src/requirements.txt
python-telegram-bot
beautifulsoup4
asyncio